Correctly calculate zoom

+++ shared/src/commonMain/kotlin/cl/emilym/sinatra/data/models/Aliases.kt
@@ -14,7 +14,7 @@ typealias Latitude = Double
 typealias Longitude = Double
 typealias Kilometer = Double
 
-typealias Pixel = Int
+typealias Pixel = Float
 typealias Radian = Double
 typealias Degree = Double
